Update index.php
Added header. Line 28 is impacting the formatting of the nav and div being of 100% width. The commenting out of this line removes the elements that allow an event to be added, modified, deleted, and otherwise halt interaction with the notes. Also removed the "Back to Overview" link due to redundancy.